+++ internal/server/server.go
@@ -1,0 +1,16 @@
+package server
+
+import "net/http"
+
+type Config struct {
+	Addr string
+}
+
+type server struct {
+	http.Server
+	Cfg Config
+}
+
+func New() {
+
+}
